chore(ui): rename "utils" to "algorithms"
It only contains functions used to search items in the timeline now \o/
