+++ crates/matrix-sdk-ui/src/timeline/algorithms.rs
